requirements.txt
Django==1.9
django-rest-swagger==0.3.4
django-simple-history==1.7.0
djangorestframework==3.3.2
psycopg2==2.6.1
python-memcached==1.57
Pillow==3.1.0
requests==2.9.1
six==1.10.0
wheel==0.24.0